way 2 add http error

+++ src/main/java/com/tw/train/restfulapi/Controller/TodoController.java
@@ -1,13 +1,11 @@
 package com.tw.train.restfulapi.Controller;
 
+import com.tw.train.restfulapi.Controller.Exception.NotFoundException;
 import com.tw.train.restfulapi.Service.TodoService;
 import com.tw.train.restfulapi.modal.Todo;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -19,7 +17,7 @@ public class TodoController {
     private TodoService todoService;
 
     @GetMapping
-    public List<Todo> getTodoList() throws IOException {
+    public List<Todo> getTodoList(){
         return todoService.getTodoList();
     }
 
@@ -34,7 +32,10 @@ public class TodoController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public List<Todo> deleteTodo(@PathVariable(value = "id") Long id){
+    public List<Todo> deleteTodo(@PathVariable(value = "id") Long id) throws NotFoundException {
+        if (todoService.getTodoById(id) == null) {
+            throw new NotFoundException();
+        }
         return todoService.deleteTodo(id);
     }
 }
